refactor(ui): split _display into display package by domain
Replace the 2k-line presentation module with focused modules and a thin compat shim.
